+++ src/main/java/org/kumoricon/service/print/BadgePrintService.java
@@ -18,6 +18,32 @@ public class BadgePrintService extends PrintService {
 
     /**
      * Prints badges for a given list of attendees to either the appropriate printer name (from
+     * the computers table), or the default printer on the server with the given offset
+     * @param attendees List of attendees
+     * @param clientIPAddress Client computer's IP address
+     * @return String Result message
+     */
+    public String printBadgesForAttendees(List<Attendee> attendees, String clientIPAddress, Integer xOffset, Integer yOffset) {
+        if (enablePrintingFromServer != null && enablePrintingFromServer) {
+            Computer client = computerService.findComputerByIP(clientIPAddress);
+            BadgePrintFormatter badgePrintFormatter =
+                    getCurrentBadgeFormatter(attendees, xOffset, yOffset);
+            try {
+                printDocument(badgePrintFormatter.getStream(), client.getPrinterName());
+            } catch (PrintException e) {
+                log.error(String.format("Error printing badge for %s: %s",
+                        clientIPAddress, e.getMessage()), e);
+                return("Error printing. No printers found? More information in server logs");
+            }
+        } else {
+            return("Printing from server not enabled. Select \"Show Badges in Browser\".");
+        }
+        return "Printed";
+    }
+
+
+    /**
+     * Prints badges for a given list of attendees to either the appropriate printer name (from
      * the computers table), or the default printer on the server.
      * @param attendees List of attendees
      * @param clientIPAddress Client computer's IP address